requirements.txt
fastapi
uvicorn
python-dotenv
faiss-cpu
pandas
numpy
langchain
openai>=1.30.0
office365-rest-python-client
unstructured
streamlit
msal
requests
pdfminer.six
PyMuPDF
pi-heif
pillow-heif
mcp
fastmcp
sentence-transformers